@@ -1,6 +1,8 @@
 module github.com/corny90/social-app-pkg
 
-go 1.20
+go 1.21
+
+toolchain go1.22.5
 
 require (
 	github.com/gocql/gocql v1.6.0
