@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.1
-// 	protoc        v3.20.3
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.27.3
 // source: grpc-proto/behaviour.proto
 
 package grpc_proto
@@ -622,7 +622,7 @@ func file_grpc_proto_behaviour_proto_rawDescGZIP() []byte {
 }
 
 var file_grpc_proto_behaviour_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
-var file_grpc_proto_behaviour_proto_goTypes = []interface{}{
+var file_grpc_proto_behaviour_proto_goTypes = []any{
 	(*PingApiBehaviourRequest)(nil),  // 0: behaviour.PingApiBehaviourRequest
 	(*PingApiBehaviourResponse)(nil), // 1: behaviour.PingApiBehaviourResponse
 	(*BehaviourFetchRequest)(nil),    // 2: behaviour.BehaviourFetchRequest
@@ -655,7 +655,7 @@ func file_grpc_proto_behaviour_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_grpc_proto_behaviour_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_behaviour_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*PingApiBehaviourRequest); i {
 			case 0:
 				return &v.state
@@ -667,7 +667,7 @@ func file_grpc_proto_behaviour_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_behaviour_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_behaviour_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*PingApiBehaviourResponse); i {
 			case 0:
 				return &v.state
@@ -679,7 +679,7 @@ func file_grpc_proto_behaviour_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_behaviour_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_behaviour_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*BehaviourFetchRequest); i {
 			case 0:
 				return &v.state
@@ -691,7 +691,7 @@ func file_grpc_proto_behaviour_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_behaviour_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_behaviour_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*BehaviourFetchResponse); i {
 			case 0:
 				return &v.state
@@ -703,7 +703,7 @@ func file_grpc_proto_behaviour_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_behaviour_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_behaviour_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*TrackRecord); i {
 			case 0:
 				return &v.state
@@ -715,7 +715,7 @@ func file_grpc_proto_behaviour_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_behaviour_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_behaviour_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*Clicks); i {
 			case 0:
 				return &v.state
@@ -727,7 +727,7 @@ func file_grpc_proto_behaviour_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_behaviour_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_behaviour_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*KeywordStats); i {
 			case 0:
 				return &v.state
@@ -739,7 +739,7 @@ func file_grpc_proto_behaviour_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_behaviour_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_behaviour_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*StatisticData); i {
 			case 0:
 				return &v.state
