@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.1
-// 	protoc        v3.20.3
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.27.3
 // source: grpc-proto/media.proto
 
 package grpc_proto
@@ -797,7 +797,7 @@ func file_grpc_proto_media_proto_rawDescGZIP() []byte {
 }
 
 var file_grpc_proto_media_proto_msgTypes = make([]protoimpl.MessageInfo, 11)
-var file_grpc_proto_media_proto_goTypes = []interface{}{
+var file_grpc_proto_media_proto_goTypes = []any{
 	(*PingApiMediaRequest)(nil),     // 0: media.PingApiMediaRequest
 	(*PingApiMediaResponse)(nil),    // 1: media.PingApiMediaResponse
 	(*PostUploadMediaRequest)(nil),  // 2: media.PostUploadMediaRequest
@@ -837,7 +837,7 @@ func file_grpc_proto_media_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_grpc_proto_media_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_media_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*PingApiMediaRequest); i {
 			case 0:
 				return &v.state
@@ -849,7 +849,7 @@ func file_grpc_proto_media_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_media_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_media_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*PingApiMediaResponse); i {
 			case 0:
 				return &v.state
@@ -861,7 +861,7 @@ func file_grpc_proto_media_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_media_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_media_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*PostUploadMediaRequest); i {
 			case 0:
 				return &v.state
@@ -873,7 +873,7 @@ func file_grpc_proto_media_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_media_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_media_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PostUploadMediaResponse); i {
 			case 0:
 				return &v.state
@@ -885,7 +885,7 @@ func file_grpc_proto_media_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_media_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_media_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PostFetchMediaRequest); i {
 			case 0:
 				return &v.state
@@ -897,7 +897,7 @@ func file_grpc_proto_media_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_media_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_media_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PostFetchMediaResponse); i {
 			case 0:
 				return &v.state
@@ -909,7 +909,7 @@ func file_grpc_proto_media_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_media_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_media_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PostsFetchMediaRequest); i {
 			case 0:
 				return &v.state
@@ -921,7 +921,7 @@ func file_grpc_proto_media_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_media_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_media_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*PostsFetchMediaResponse); i {
 			case 0:
 				return &v.state
@@ -933,7 +933,7 @@ func file_grpc_proto_media_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_media_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_media_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*PostDeleteMediaRequest); i {
 			case 0:
 				return &v.state
@@ -945,7 +945,7 @@ func file_grpc_proto_media_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_media_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_media_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*PostDeleteMediaResponse); i {
 			case 0:
 				return &v.state
@@ -957,7 +957,7 @@ func file_grpc_proto_media_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_media_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_media_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*Media); i {
 			case 0:
 				return &v.state
