@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.1
-// 	protoc        v3.20.3
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.27.3
 // source: grpc-proto/message.proto
 
 package grpc_proto
@@ -1409,6 +1409,172 @@ func (x *CompleteTaskMessageResponse) GetStatus() string {
 	return ""
 }
 
+type ConversationRender struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	ConversationID string         `protobuf:"bytes,1,opt,name=conversationID,proto3" json:"conversationID,omitempty"`
+	Participants   []*Participant `protobuf:"bytes,2,rep,name=participants,proto3" json:"participants,omitempty"`
+	CreatedAt      string         `protobuf:"bytes,3,opt,name=createdAt,proto3" json:"createdAt,omitempty"`
+	EditedAt       string         `protobuf:"bytes,4,opt,name=editedAt,proto3" json:"editedAt,omitempty"`
+	Metadata       string         `protobuf:"bytes,5,opt,name=metadata,proto3" json:"metadata,omitempty"`
+	State          string         `protobuf:"bytes,6,opt,name=state,proto3" json:"state,omitempty"`
+}
+
+func (x *ConversationRender) Reset() {
+	*x = ConversationRender{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_grpc_proto_message_proto_msgTypes[26]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *ConversationRender) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ConversationRender) ProtoMessage() {}
+
+func (x *ConversationRender) ProtoReflect() protoreflect.Message {
+	mi := &file_grpc_proto_message_proto_msgTypes[26]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ConversationRender.ProtoReflect.Descriptor instead.
+func (*ConversationRender) Descriptor() ([]byte, []int) {
+	return file_grpc_proto_message_proto_rawDescGZIP(), []int{26}
+}
+
+func (x *ConversationRender) GetConversationID() string {
+	if x != nil {
+		return x.ConversationID
+	}
+	return ""
+}
+
+func (x *ConversationRender) GetParticipants() []*Participant {
+	if x != nil {
+		return x.Participants
+	}
+	return nil
+}
+
+func (x *ConversationRender) GetCreatedAt() string {
+	if x != nil {
+		return x.CreatedAt
+	}
+	return ""
+}
+
+func (x *ConversationRender) GetEditedAt() string {
+	if x != nil {
+		return x.EditedAt
+	}
+	return ""
+}
+
+func (x *ConversationRender) GetMetadata() string {
+	if x != nil {
+		return x.Metadata
+	}
+	return ""
+}
+
+func (x *ConversationRender) GetState() string {
+	if x != nil {
+		return x.State
+	}
+	return ""
+}
+
+type Participant struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	UserID    string `protobuf:"bytes,1,opt,name=userID,proto3" json:"userID,omitempty"`
+	UserType  string `protobuf:"bytes,2,opt,name=UserType,proto3" json:"UserType,omitempty"`
+	UserRole  string `protobuf:"bytes,3,opt,name=UserRole,proto3" json:"UserRole,omitempty"`
+	Username  string `protobuf:"bytes,4,opt,name=Username,proto3" json:"Username,omitempty"`
+	AvatarUrl string `protobuf:"bytes,5,opt,name=AvatarUrl,proto3" json:"AvatarUrl,omitempty"`
+}
+
+func (x *Participant) Reset() {
+	*x = Participant{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_grpc_proto_message_proto_msgTypes[27]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *Participant) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*Participant) ProtoMessage() {}
+
+func (x *Participant) ProtoReflect() protoreflect.Message {
+	mi := &file_grpc_proto_message_proto_msgTypes[27]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use Participant.ProtoReflect.Descriptor instead.
+func (*Participant) Descriptor() ([]byte, []int) {
+	return file_grpc_proto_message_proto_rawDescGZIP(), []int{27}
+}
+
+func (x *Participant) GetUserID() string {
+	if x != nil {
+		return x.UserID
+	}
+	return ""
+}
+
+func (x *Participant) GetUserType() string {
+	if x != nil {
+		return x.UserType
+	}
+	return ""
+}
+
+func (x *Participant) GetUserRole() string {
+	if x != nil {
+		return x.UserRole
+	}
+	return ""
+}
+
+func (x *Participant) GetUsername() string {
+	if x != nil {
+		return x.Username
+	}
+	return ""
+}
+
+func (x *Participant) GetAvatarUrl() string {
+	if x != nil {
+		return x.AvatarUrl
+	}
+	return ""
+}
+
 // Conversation
 type Conversation struct {
 	state         protoimpl.MessageState
@@ -1426,7 +1592,7 @@ type Conversation struct {
 func (x *Conversation) Reset() {
 	*x = Conversation{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_grpc_proto_message_proto_msgTypes[26]
+		mi := &file_grpc_proto_message_proto_msgTypes[28]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1439,7 +1605,7 @@ func (x *Conversation) String() string {
 func (*Conversation) ProtoMessage() {}
 
 func (x *Conversation) ProtoReflect() protoreflect.Message {
-	mi := &file_grpc_proto_message_proto_msgTypes[26]
+	mi := &file_grpc_proto_message_proto_msgTypes[28]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -1452,7 +1618,7 @@ func (x *Conversation) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use Conversation.ProtoReflect.Descriptor instead.
 func (*Conversation) Descriptor() ([]byte, []int) {
-	return file_grpc_proto_message_proto_rawDescGZIP(), []int{26}
+	return file_grpc_proto_message_proto_rawDescGZIP(), []int{28}
 }
 
 func (x *Conversation) GetConversationID() string {
@@ -1516,7 +1682,7 @@ type Message struct {
 func (x *Message) Reset() {
 	*x = Message{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_grpc_proto_message_proto_msgTypes[27]
+		mi := &file_grpc_proto_message_proto_msgTypes[29]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1529,7 +1695,7 @@ func (x *Message) String() string {
 func (*Message) ProtoMessage() {}
 
 func (x *Message) ProtoReflect() protoreflect.Message {
-	mi := &file_grpc_proto_message_proto_msgTypes[27]
+	mi := &file_grpc_proto_message_proto_msgTypes[29]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -1542,7 +1708,7 @@ func (x *Message) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use Message.ProtoReflect.Descriptor instead.
 func (*Message) Descriptor() ([]byte, []int) {
-	return file_grpc_proto_message_proto_rawDescGZIP(), []int{27}
+	return file_grpc_proto_message_proto_rawDescGZIP(), []int{29}
 }
 
 func (x *Message) GetMessageID() string {
@@ -1620,7 +1786,7 @@ type TaskData struct {
 func (x *TaskData) Reset() {
 	*x = TaskData{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_grpc_proto_message_proto_msgTypes[28]
+		mi := &file_grpc_proto_message_proto_msgTypes[30]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -1633,7 +1799,7 @@ func (x *TaskData) String() string {
 func (*TaskData) ProtoMessage() {}
 
 func (x *TaskData) ProtoReflect() protoreflect.Message {
-	mi := &file_grpc_proto_message_proto_msgTypes[28]
+	mi := &file_grpc_proto_message_proto_msgTypes[30]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -1646,7 +1812,7 @@ func (x *TaskData) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use TaskData.ProtoReflect.Descriptor instead.
 func (*TaskData) Descriptor() ([]byte, []int) {
-	return file_grpc_proto_message_proto_rawDescGZIP(), []int{28}
+	return file_grpc_proto_message_proto_rawDescGZIP(), []int{30}
 }
 
 func (x *TaskData) GetMessageID() string {
@@ -1848,126 +2014,150 @@ var file_grpc_proto_message_proto_rawDesc = []byte{
 	0x61, 0x74, 0x61, 0x22, 0x35, 0x0a, 0x1b, 0x43, 0x6f, 0x6d, 0x70, 0x6c, 0x65, 0x74, 0x65, 0x54,
 	0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
 	0x73, 0x65, 0x12, 0x16, 0x0a, 0x06, 0x73, 0x74, 0x61, 0x74, 0x75, 0x73, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x06, 0x73, 0x74, 0x61, 0x74, 0x75, 0x73, 0x22, 0xc6, 0x01, 0x0a, 0x0c, 0x43,
-	0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x26, 0x0a, 0x0e, 0x63,
-	0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x44, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x0e, 0x63, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f,
-	0x6e, 0x49, 0x44, 0x12, 0x22, 0x0a, 0x0c, 0x70, 0x61, 0x72, 0x74, 0x69, 0x63, 0x69, 0x70, 0x61,
-	0x6e, 0x74, 0x73, 0x18, 0x02, 0x20, 0x03, 0x28, 0x09, 0x52, 0x0c, 0x70, 0x61, 0x72, 0x74, 0x69,
-	0x63, 0x69, 0x70, 0x61, 0x6e, 0x74, 0x73, 0x12, 0x1c, 0x0a, 0x09, 0x63, 0x72, 0x65, 0x61, 0x74,
-	0x65, 0x64, 0x41, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x63, 0x72, 0x65, 0x61,
-	0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x1a, 0x0a, 0x08, 0x65, 0x64, 0x69, 0x74, 0x65, 0x64, 0x41,
-	0x74, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x65, 0x64, 0x69, 0x74, 0x65, 0x64, 0x41,
-	0x74, 0x12, 0x1a, 0x0a, 0x08, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x18, 0x05, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x08, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x12, 0x14, 0x0a,
-	0x05, 0x73, 0x74, 0x61, 0x74, 0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x73, 0x74,
-	0x61, 0x74, 0x65, 0x22, 0xf1, 0x01, 0x0a, 0x07, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12,
-	0x1c, 0x0a, 0x09, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49, 0x44, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x09, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49, 0x44, 0x12, 0x26, 0x0a,
-	0x0e, 0x63, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x44, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0e, 0x63, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74,
-	0x69, 0x6f, 0x6e, 0x49, 0x44, 0x12, 0x16, 0x0a, 0x06, 0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x18,
-	0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x12, 0x1e, 0x0a,
-	0x0a, 0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x54, 0x79, 0x70, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28,
-	0x09, 0x52, 0x0a, 0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x54, 0x79, 0x70, 0x65, 0x12, 0x18, 0x0a,
-	0x07, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07,
-	0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x74, 0x61, 0x74, 0x65,
-	0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x73, 0x74, 0x61, 0x74, 0x65, 0x12, 0x1c, 0x0a,
-	0x09, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x18, 0x07, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x09, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x1a, 0x0a, 0x08, 0x6d,
-	0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x18, 0x08, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x6d,
-	0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x22, 0xf0, 0x01, 0x0a, 0x08, 0x54, 0x61, 0x73, 0x6b,
-	0x44, 0x61, 0x74, 0x61, 0x12, 0x1c, 0x0a, 0x09, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49,
-	0x44, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
-	0x49, 0x44, 0x12, 0x26, 0x0a, 0x0e, 0x63, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69,
-	0x6f, 0x6e, 0x49, 0x44, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0e, 0x63, 0x6f, 0x6e, 0x76,
-	0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x44, 0x12, 0x16, 0x0a, 0x06, 0x73, 0x65,
-	0x6e, 0x64, 0x65, 0x72, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x73, 0x65, 0x6e, 0x64,
-	0x65, 0x72, 0x12, 0x18, 0x0a, 0x07, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x18, 0x04, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x07, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x12, 0x1c, 0x0a, 0x09,
-	0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x18, 0x05, 0x20, 0x01, 0x28, 0x03, 0x52,
-	0x09, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x74,
-	0x61, 0x74, 0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x73, 0x74, 0x61, 0x74, 0x65,
-	0x12, 0x1a, 0x0a, 0x08, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x18, 0x07, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x08, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x12, 0x1c, 0x0a, 0x09,
-	0x72, 0x65, 0x63, 0x69, 0x70, 0x69, 0x65, 0x6e, 0x74, 0x18, 0x08, 0x20, 0x03, 0x28, 0x09, 0x52,
-	0x09, 0x72, 0x65, 0x63, 0x69, 0x70, 0x69, 0x65, 0x6e, 0x74, 0x32, 0xa1, 0x09, 0x0a, 0x0e, 0x4d,
-	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x51, 0x0a,
-	0x0e, 0x50, 0x69, 0x6e, 0x67, 0x41, 0x70, 0x69, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12,
-	0x1e, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x50, 0x69, 0x6e, 0x67, 0x41, 0x70,
-	0x69, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
-	0x1f, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x50, 0x69, 0x6e, 0x67, 0x41, 0x70,
-	0x69, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
-	0x12, 0x54, 0x0a, 0x0f, 0x43, 0x75, 0x73, 0x74, 0x6f, 0x6d, 0x65, 0x72, 0x4d, 0x65, 0x73, 0x73,
-	0x61, 0x67, 0x65, 0x12, 0x1f, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43, 0x75,
-	0x73, 0x74, 0x6f, 0x6d, 0x65, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x1a, 0x20, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43,
-	0x75, 0x73, 0x74, 0x6f, 0x6d, 0x65, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65,
-	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x51, 0x0a, 0x0e, 0x43, 0x68, 0x61, 0x74, 0x62, 0x6f,
-	0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1e, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x2e, 0x43, 0x68, 0x61, 0x74, 0x62, 0x6f, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67,
-	0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1f, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x2e, 0x43, 0x68, 0x61, 0x74, 0x62, 0x6f, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67,
-	0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x51, 0x0a, 0x0e, 0x50, 0x69, 0x6e,
-	0x67, 0x54, 0x61, 0x73, 0x6b, 0x73, 0x51, 0x75, 0x65, 0x75, 0x65, 0x12, 0x1e, 0x2e, 0x6d, 0x65,
-	0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x50, 0x69, 0x6e, 0x67, 0x54, 0x61, 0x73, 0x6b, 0x73, 0x51,
-	0x75, 0x65, 0x75, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1f, 0x2e, 0x6d, 0x65,
-	0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x50, 0x69, 0x6e, 0x67, 0x54, 0x61, 0x73, 0x6b, 0x73, 0x51,
-	0x75, 0x65, 0x75, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x5a, 0x0a, 0x11,
-	0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x54, 0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67,
-	0x65, 0x12, 0x21, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43, 0x72, 0x65, 0x61,
-	0x74, 0x65, 0x54, 0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x1a, 0x22, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43,
+	0x28, 0x09, 0x52, 0x06, 0x73, 0x74, 0x61, 0x74, 0x75, 0x73, 0x22, 0xe2, 0x01, 0x0a, 0x12, 0x43,
+	0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x6e, 0x64, 0x65,
+	0x72, 0x12, 0x26, 0x0a, 0x0e, 0x63, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f,
+	0x6e, 0x49, 0x44, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0e, 0x63, 0x6f, 0x6e, 0x76, 0x65,
+	0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x44, 0x12, 0x38, 0x0a, 0x0c, 0x70, 0x61, 0x72,
+	0x74, 0x69, 0x63, 0x69, 0x70, 0x61, 0x6e, 0x74, 0x73, 0x18, 0x02, 0x20, 0x03, 0x28, 0x0b, 0x32,
+	0x14, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x50, 0x61, 0x72, 0x74, 0x69, 0x63,
+	0x69, 0x70, 0x61, 0x6e, 0x74, 0x52, 0x0c, 0x70, 0x61, 0x72, 0x74, 0x69, 0x63, 0x69, 0x70, 0x61,
+	0x6e, 0x74, 0x73, 0x12, 0x1c, 0x0a, 0x09, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74,
+	0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41,
+	0x74, 0x12, 0x1a, 0x0a, 0x08, 0x65, 0x64, 0x69, 0x74, 0x65, 0x64, 0x41, 0x74, 0x18, 0x04, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x08, 0x65, 0x64, 0x69, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x1a, 0x0a,
+	0x08, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x08, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x74, 0x61,
+	0x74, 0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x73, 0x74, 0x61, 0x74, 0x65, 0x22,
+	0x97, 0x01, 0x0a, 0x0b, 0x50, 0x61, 0x72, 0x74, 0x69, 0x63, 0x69, 0x70, 0x61, 0x6e, 0x74, 0x12,
+	0x16, 0x0a, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x44, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x44, 0x12, 0x1a, 0x0a, 0x08, 0x55, 0x73, 0x65, 0x72, 0x54,
+	0x79, 0x70, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x55, 0x73, 0x65, 0x72, 0x54,
+	0x79, 0x70, 0x65, 0x12, 0x1a, 0x0a, 0x08, 0x55, 0x73, 0x65, 0x72, 0x52, 0x6f, 0x6c, 0x65, 0x18,
+	0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x55, 0x73, 0x65, 0x72, 0x52, 0x6f, 0x6c, 0x65, 0x12,
+	0x1a, 0x0a, 0x08, 0x55, 0x73, 0x65, 0x72, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28,
+	0x09, 0x52, 0x08, 0x55, 0x73, 0x65, 0x72, 0x6e, 0x61, 0x6d, 0x65, 0x12, 0x1c, 0x0a, 0x09, 0x41,
+	0x76, 0x61, 0x74, 0x61, 0x72, 0x55, 0x72, 0x6c, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09,
+	0x41, 0x76, 0x61, 0x74, 0x61, 0x72, 0x55, 0x72, 0x6c, 0x22, 0xc6, 0x01, 0x0a, 0x0c, 0x43, 0x6f,
+	0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x26, 0x0a, 0x0e, 0x63, 0x6f,
+	0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x44, 0x18, 0x01, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x0e, 0x63, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e,
+	0x49, 0x44, 0x12, 0x22, 0x0a, 0x0c, 0x70, 0x61, 0x72, 0x74, 0x69, 0x63, 0x69, 0x70, 0x61, 0x6e,
+	0x74, 0x73, 0x18, 0x02, 0x20, 0x03, 0x28, 0x09, 0x52, 0x0c, 0x70, 0x61, 0x72, 0x74, 0x69, 0x63,
+	0x69, 0x70, 0x61, 0x6e, 0x74, 0x73, 0x12, 0x1c, 0x0a, 0x09, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65,
+	0x64, 0x41, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x63, 0x72, 0x65, 0x61, 0x74,
+	0x65, 0x64, 0x41, 0x74, 0x12, 0x1a, 0x0a, 0x08, 0x65, 0x64, 0x69, 0x74, 0x65, 0x64, 0x41, 0x74,
+	0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x65, 0x64, 0x69, 0x74, 0x65, 0x64, 0x41, 0x74,
+	0x12, 0x1a, 0x0a, 0x08, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x18, 0x05, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x08, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x12, 0x14, 0x0a, 0x05,
+	0x73, 0x74, 0x61, 0x74, 0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x73, 0x74, 0x61,
+	0x74, 0x65, 0x22, 0xf1, 0x01, 0x0a, 0x07, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1c,
+	0x0a, 0x09, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49, 0x44, 0x18, 0x01, 0x20, 0x01, 0x28,
+	0x09, 0x52, 0x09, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49, 0x44, 0x12, 0x26, 0x0a, 0x0e,
+	0x63, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x44, 0x18, 0x02,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x0e, 0x63, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69,
+	0x6f, 0x6e, 0x49, 0x44, 0x12, 0x16, 0x0a, 0x06, 0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x18, 0x03,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x12, 0x1e, 0x0a, 0x0a,
+	0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x54, 0x79, 0x70, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x0a, 0x73, 0x65, 0x6e, 0x64, 0x65, 0x72, 0x54, 0x79, 0x70, 0x65, 0x12, 0x18, 0x0a, 0x07,
+	0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x63,
+	0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x74, 0x61, 0x74, 0x65, 0x18,
+	0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x73, 0x74, 0x61, 0x74, 0x65, 0x12, 0x1c, 0x0a, 0x09,
+	0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x18, 0x07, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x09, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x1a, 0x0a, 0x08, 0x6d, 0x65,
+	0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x18, 0x08, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x6d, 0x65,
+	0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x22, 0xf0, 0x01, 0x0a, 0x08, 0x54, 0x61, 0x73, 0x6b, 0x44,
+	0x61, 0x74, 0x61, 0x12, 0x1c, 0x0a, 0x09, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49, 0x44,
+	0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x49,
+	0x44, 0x12, 0x26, 0x0a, 0x0e, 0x63, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f,
+	0x6e, 0x49, 0x44, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0e, 0x63, 0x6f, 0x6e, 0x76, 0x65,
+	0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x44, 0x12, 0x16, 0x0a, 0x06, 0x73, 0x65, 0x6e,
+	0x64, 0x65, 0x72, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x73, 0x65, 0x6e, 0x64, 0x65,
+	0x72, 0x12, 0x18, 0x0a, 0x07, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x18, 0x04, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x07, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x12, 0x1c, 0x0a, 0x09, 0x74,
+	0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x18, 0x05, 0x20, 0x01, 0x28, 0x03, 0x52, 0x09,
+	0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x74, 0x61,
+	0x74, 0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x73, 0x74, 0x61, 0x74, 0x65, 0x12,
+	0x1a, 0x0a, 0x08, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x18, 0x07, 0x20, 0x01, 0x28,
+	0x09, 0x52, 0x08, 0x6d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x12, 0x1c, 0x0a, 0x09, 0x72,
+	0x65, 0x63, 0x69, 0x70, 0x69, 0x65, 0x6e, 0x74, 0x18, 0x08, 0x20, 0x03, 0x28, 0x09, 0x52, 0x09,
+	0x72, 0x65, 0x63, 0x69, 0x70, 0x69, 0x65, 0x6e, 0x74, 0x32, 0xa1, 0x09, 0x0a, 0x0e, 0x4d, 0x65,
+	0x73, 0x73, 0x61, 0x67, 0x65, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x51, 0x0a, 0x0e,
+	0x50, 0x69, 0x6e, 0x67, 0x41, 0x70, 0x69, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1e,
+	0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x50, 0x69, 0x6e, 0x67, 0x41, 0x70, 0x69,
+	0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1f,
+	0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x50, 0x69, 0x6e, 0x67, 0x41, 0x70, 0x69,
+	0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12,
+	0x54, 0x0a, 0x0f, 0x43, 0x75, 0x73, 0x74, 0x6f, 0x6d, 0x65, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61,
+	0x67, 0x65, 0x12, 0x1f, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43, 0x75, 0x73,
+	0x74, 0x6f, 0x6d, 0x65, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x75,
+	0x65, 0x73, 0x74, 0x1a, 0x20, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43, 0x75,
+	0x73, 0x74, 0x6f, 0x6d, 0x65, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x73,
+	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x51, 0x0a, 0x0e, 0x43, 0x68, 0x61, 0x74, 0x62, 0x6f, 0x74,
+	0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1e, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67,
+	0x65, 0x2e, 0x43, 0x68, 0x61, 0x74, 0x62, 0x6f, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1f, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67,
+	0x65, 0x2e, 0x43, 0x68, 0x61, 0x74, 0x62, 0x6f, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x51, 0x0a, 0x0e, 0x50, 0x69, 0x6e, 0x67,
+	0x54, 0x61, 0x73, 0x6b, 0x73, 0x51, 0x75, 0x65, 0x75, 0x65, 0x12, 0x1e, 0x2e, 0x6d, 0x65, 0x73,
+	0x73, 0x61, 0x67, 0x65, 0x2e, 0x50, 0x69, 0x6e, 0x67, 0x54, 0x61, 0x73, 0x6b, 0x73, 0x51, 0x75,
+	0x65, 0x75, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1f, 0x2e, 0x6d, 0x65, 0x73,
+	0x73, 0x61, 0x67, 0x65, 0x2e, 0x50, 0x69, 0x6e, 0x67, 0x54, 0x61, 0x73, 0x6b, 0x73, 0x51, 0x75,
+	0x65, 0x75, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x5a, 0x0a, 0x11, 0x43,
 	0x72, 0x65, 0x61, 0x74, 0x65, 0x54, 0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
-	0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x60, 0x0a, 0x13, 0x43, 0x6f, 0x6d, 0x70,
-	0x6c, 0x65, 0x74, 0x65, 0x54, 0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12,
-	0x23, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43, 0x6f, 0x6d, 0x70, 0x6c, 0x65,
-	0x74, 0x65, 0x54, 0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x1a, 0x24, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43,
-	0x6f, 0x6d, 0x70, 0x6c, 0x65, 0x74, 0x65, 0x54, 0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x54, 0x0a, 0x0f, 0x55, 0x73,
-	0x65, 0x72, 0x53, 0x65, 0x6e, 0x64, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1f, 0x2e,
+	0x12, 0x21, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43, 0x72, 0x65, 0x61, 0x74,
+	0x65, 0x54, 0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x75,
+	0x65, 0x73, 0x74, 0x1a, 0x22, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43, 0x72,
+	0x65, 0x61, 0x74, 0x65, 0x54, 0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52,
+	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x60, 0x0a, 0x13, 0x43, 0x6f, 0x6d, 0x70, 0x6c,
+	0x65, 0x74, 0x65, 0x54, 0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x23,
+	0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43, 0x6f, 0x6d, 0x70, 0x6c, 0x65, 0x74,
+	0x65, 0x54, 0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x75,
+	0x65, 0x73, 0x74, 0x1a, 0x24, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x43, 0x6f,
+	0x6d, 0x70, 0x6c, 0x65, 0x74, 0x65, 0x54, 0x61, 0x73, 0x6b, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67,
+	0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x54, 0x0a, 0x0f, 0x55, 0x73, 0x65,
+	0x72, 0x53, 0x65, 0x6e, 0x64, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1f, 0x2e, 0x6d,
+	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x55, 0x73, 0x65, 0x72, 0x53, 0x65, 0x6e, 0x64, 0x4d,
+	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x20, 0x2e,
 	0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x55, 0x73, 0x65, 0x72, 0x53, 0x65, 0x6e, 0x64,
-	0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x20,
-	0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x55, 0x73, 0x65, 0x72, 0x53, 0x65, 0x6e,
-	0x64, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
-	0x12, 0x48, 0x0a, 0x0b, 0x45, 0x64, 0x69, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12,
-	0x1b, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x45, 0x64, 0x69, 0x74, 0x4d, 0x65,
-	0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1c, 0x2e, 0x6d,
-	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x45, 0x64, 0x69, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x4b, 0x0a, 0x0c, 0x46, 0x65,
-	0x74, 0x63, 0x68, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1c, 0x2e, 0x6d, 0x65, 0x73,
-	0x73, 0x61, 0x67, 0x65, 0x2e, 0x46, 0x65, 0x74, 0x63, 0x68, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67,
-	0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1d, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x2e, 0x46, 0x65, 0x74, 0x63, 0x68, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x4e, 0x0a, 0x0d, 0x44, 0x65, 0x6c, 0x65, 0x74,
-	0x65, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1d, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x2e, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
-	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1e, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67,
+	0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12,
+	0x48, 0x0a, 0x0b, 0x45, 0x64, 0x69, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1b,
+	0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x45, 0x64, 0x69, 0x74, 0x4d, 0x65, 0x73,
+	0x73, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1c, 0x2e, 0x6d, 0x65,
+	0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x45, 0x64, 0x69, 0x74, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67,
+	0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x4b, 0x0a, 0x0c, 0x46, 0x65, 0x74,
+	0x63, 0x68, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1c, 0x2e, 0x6d, 0x65, 0x73, 0x73,
+	0x61, 0x67, 0x65, 0x2e, 0x46, 0x65, 0x74, 0x63, 0x68, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1d, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67,
+	0x65, 0x2e, 0x46, 0x65, 0x74, 0x63, 0x68, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65,
+	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x4e, 0x0a, 0x0d, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65,
+	0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x1d, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67,
 	0x65, 0x2e, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x72, 0x0a, 0x19, 0x46, 0x65, 0x74, 0x63, 0x68,
-	0x43, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x4d, 0x65, 0x73, 0x73,
-	0x61, 0x67, 0x65, 0x73, 0x12, 0x29, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x46,
-	0x65, 0x74, 0x63, 0x68, 0x43, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e,
-	0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
-	0x2a, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x46, 0x65, 0x74, 0x63, 0x68, 0x43,
+	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1e, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x2e, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x52, 0x65,
+	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x72, 0x0a, 0x19, 0x46, 0x65, 0x74, 0x63, 0x68, 0x43,
 	0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x4d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x5d, 0x0a, 0x12, 0x44,
+	0x67, 0x65, 0x73, 0x12, 0x29, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x46, 0x65,
+	0x74, 0x63, 0x68, 0x43, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x4d,
+	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x2a,
+	0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x46, 0x65, 0x74, 0x63, 0x68, 0x43, 0x6f,
+	0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67,
+	0x65, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x5d, 0x0a, 0x12, 0x44, 0x65,
+	0x6c, 0x65, 0x74, 0x65, 0x43, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e,
+	0x12, 0x22, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x44, 0x65, 0x6c, 0x65, 0x74,
+	0x65, 0x43, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x71,
+	0x75, 0x65, 0x73, 0x74, 0x1a, 0x23, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x44,
 	0x65, 0x6c, 0x65, 0x74, 0x65, 0x43, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f,
-	0x6e, 0x12, 0x22, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x44, 0x65, 0x6c, 0x65,
-	0x74, 0x65, 0x43, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x23, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e,
-	0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x43, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69,
-	0x6f, 0x6e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x72, 0x0a, 0x19, 0x46, 0x65,
-	0x74, 0x63, 0x68, 0x41, 0x6c, 0x6c, 0x55, 0x73, 0x65, 0x72, 0x43, 0x6f, 0x6e, 0x76, 0x65, 0x72,
-	0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x12, 0x29, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67,
-	0x65, 0x2e, 0x46, 0x65, 0x74, 0x63, 0x68, 0x41, 0x6c, 0x6c, 0x55, 0x73, 0x65, 0x72, 0x43, 0x6f,
-	0x6e, 0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65,
-	0x73, 0x74, 0x1a, 0x2a, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x46, 0x65, 0x74,
+	0x6e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x72, 0x0a, 0x19, 0x46, 0x65, 0x74,
 	0x63, 0x68, 0x41, 0x6c, 0x6c, 0x55, 0x73, 0x65, 0x72, 0x43, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73,
-	0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x42, 0x18,
-	0x5a, 0x16, 0x2f, 0x67, 0x72, 0x70, 0x63, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x3b, 0x67, 0x72,
-	0x70, 0x63, 0x5f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x12, 0x29, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65,
+	0x2e, 0x46, 0x65, 0x74, 0x63, 0x68, 0x41, 0x6c, 0x6c, 0x55, 0x73, 0x65, 0x72, 0x43, 0x6f, 0x6e,
+	0x76, 0x65, 0x72, 0x73, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
+	0x74, 0x1a, 0x2a, 0x2e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x2e, 0x46, 0x65, 0x74, 0x63,
+	0x68, 0x41, 0x6c, 0x6c, 0x55, 0x73, 0x65, 0x72, 0x43, 0x6f, 0x6e, 0x76, 0x65, 0x72, 0x73, 0x61,
+	0x74, 0x69, 0x6f, 0x6e, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x42, 0x18, 0x5a,
+	0x16, 0x2f, 0x67, 0x72, 0x70, 0x63, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x3b, 0x67, 0x72, 0x70,
+	0x63, 0x5f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -1982,8 +2172,8 @@ func file_grpc_proto_message_proto_rawDescGZIP() []byte {
 	return file_grpc_proto_message_proto_rawDescData
 }
 
-var file_grpc_proto_message_proto_msgTypes = make([]protoimpl.MessageInfo, 29)
-var file_grpc_proto_message_proto_goTypes = []interface{}{
+var file_grpc_proto_message_proto_msgTypes = make([]protoimpl.MessageInfo, 31)
+var file_grpc_proto_message_proto_goTypes = []any{
 	(*PingApiMessageRequest)(nil),             // 0: message.PingApiMessageRequest
 	(*PingApiMessageResponse)(nil),            // 1: message.PingApiMessageResponse
 	(*PingTasksQueueRequest)(nil),             // 2: message.PingTasksQueueRequest
@@ -2010,51 +2200,54 @@ var file_grpc_proto_message_proto_goTypes = []interface{}{
 	(*CreateTaskMessageResponse)(nil),         // 23: message.CreateTaskMessageResponse
 	(*CompleteTaskMessageRequest)(nil),        // 24: message.CompleteTaskMessageRequest
 	(*CompleteTaskMessageResponse)(nil),       // 25: message.CompleteTaskMessageResponse
-	(*Conversation)(nil),                      // 26: message.Conversation
-	(*Message)(nil),                           // 27: message.Message
-	(*TaskData)(nil),                          // 28: message.TaskData
+	(*ConversationRender)(nil),                // 26: message.ConversationRender
+	(*Participant)(nil),                       // 27: message.Participant
+	(*Conversation)(nil),                      // 28: message.Conversation
+	(*Message)(nil),                           // 29: message.Message
+	(*TaskData)(nil),                          // 30: message.TaskData
 }
 var file_grpc_proto_message_proto_depIdxs = []int32{
-	27, // 0: message.CustomerMessageResponse.message:type_name -> message.Message
-	27, // 1: message.UserSendMessageResponse.userMessage:type_name -> message.Message
-	27, // 2: message.UserSendMessageResponse.botMessage:type_name -> message.Message
-	27, // 3: message.FetchMessageResponse.message:type_name -> message.Message
-	27, // 4: message.FetchConversationMessagesResponse.messages:type_name -> message.Message
-	26, // 5: message.FetchAllUserConversationsResponse.conversations:type_name -> message.Conversation
-	27, // 6: message.ChatbotMessageRequest.message:type_name -> message.Message
-	28, // 7: message.CreateTaskMessageRequest.taskData:type_name -> message.TaskData
-	28, // 8: message.CompleteTaskMessageRequest.taskData:type_name -> message.TaskData
-	0,  // 9: message.MessageService.PingApiMessage:input_type -> message.PingApiMessageRequest
-	4,  // 10: message.MessageService.CustomerMessage:input_type -> message.CustomerMessageRequest
-	20, // 11: message.MessageService.ChatbotMessage:input_type -> message.ChatbotMessageRequest
-	2,  // 12: message.MessageService.PingTasksQueue:input_type -> message.PingTasksQueueRequest
-	22, // 13: message.MessageService.CreateTaskMessage:input_type -> message.CreateTaskMessageRequest
-	24, // 14: message.MessageService.CompleteTaskMessage:input_type -> message.CompleteTaskMessageRequest
-	6,  // 15: message.MessageService.UserSendMessage:input_type -> message.UserSendMessageRequest
-	8,  // 16: message.MessageService.EditMessage:input_type -> message.EditMessageRequest
-	10, // 17: message.MessageService.FetchMessage:input_type -> message.FetchMessageRequest
-	12, // 18: message.MessageService.DeleteMessage:input_type -> message.DeleteMessageRequest
-	14, // 19: message.MessageService.FetchConversationMessages:input_type -> message.FetchConversationMessagesRequest
-	16, // 20: message.MessageService.DeleteConversation:input_type -> message.DeleteConversationRequest
-	18, // 21: message.MessageService.FetchAllUserConversations:input_type -> message.FetchAllUserConversationsRequest
-	1,  // 22: message.MessageService.PingApiMessage:output_type -> message.PingApiMessageResponse
-	5,  // 23: message.MessageService.CustomerMessage:output_type -> message.CustomerMessageResponse
-	21, // 24: message.MessageService.ChatbotMessage:output_type -> message.ChatbotMessageResponse
-	3,  // 25: message.MessageService.PingTasksQueue:output_type -> message.PingTasksQueueResponse
-	23, // 26: message.MessageService.CreateTaskMessage:output_type -> message.CreateTaskMessageResponse
-	25, // 27: message.MessageService.CompleteTaskMessage:output_type -> message.CompleteTaskMessageResponse
-	7,  // 28: message.MessageService.UserSendMessage:output_type -> message.UserSendMessageResponse
-	9,  // 29: message.MessageService.EditMessage:output_type -> message.EditMessageResponse
-	11, // 30: message.MessageService.FetchMessage:output_type -> message.FetchMessageResponse
-	13, // 31: message.MessageService.DeleteMessage:output_type -> message.DeleteMessageResponse
-	15, // 32: message.MessageService.FetchConversationMessages:output_type -> message.FetchConversationMessagesResponse
-	17, // 33: message.MessageService.DeleteConversation:output_type -> message.DeleteConversationResponse
-	19, // 34: message.MessageService.FetchAllUserConversations:output_type -> message.FetchAllUserConversationsResponse
-	22, // [22:35] is the sub-list for method output_type
-	9,  // [9:22] is the sub-list for method input_type
-	9,  // [9:9] is the sub-list for extension type_name
-	9,  // [9:9] is the sub-list for extension extendee
-	0,  // [0:9] is the sub-list for field type_name
+	29, // 0: message.CustomerMessageResponse.message:type_name -> message.Message
+	29, // 1: message.UserSendMessageResponse.userMessage:type_name -> message.Message
+	29, // 2: message.UserSendMessageResponse.botMessage:type_name -> message.Message
+	29, // 3: message.FetchMessageResponse.message:type_name -> message.Message
+	29, // 4: message.FetchConversationMessagesResponse.messages:type_name -> message.Message
+	28, // 5: message.FetchAllUserConversationsResponse.conversations:type_name -> message.Conversation
+	29, // 6: message.ChatbotMessageRequest.message:type_name -> message.Message
+	30, // 7: message.CreateTaskMessageRequest.taskData:type_name -> message.TaskData
+	30, // 8: message.CompleteTaskMessageRequest.taskData:type_name -> message.TaskData
+	27, // 9: message.ConversationRender.participants:type_name -> message.Participant
+	0,  // 10: message.MessageService.PingApiMessage:input_type -> message.PingApiMessageRequest
+	4,  // 11: message.MessageService.CustomerMessage:input_type -> message.CustomerMessageRequest
+	20, // 12: message.MessageService.ChatbotMessage:input_type -> message.ChatbotMessageRequest
+	2,  // 13: message.MessageService.PingTasksQueue:input_type -> message.PingTasksQueueRequest
+	22, // 14: message.MessageService.CreateTaskMessage:input_type -> message.CreateTaskMessageRequest
+	24, // 15: message.MessageService.CompleteTaskMessage:input_type -> message.CompleteTaskMessageRequest
+	6,  // 16: message.MessageService.UserSendMessage:input_type -> message.UserSendMessageRequest
+	8,  // 17: message.MessageService.EditMessage:input_type -> message.EditMessageRequest
+	10, // 18: message.MessageService.FetchMessage:input_type -> message.FetchMessageRequest
+	12, // 19: message.MessageService.DeleteMessage:input_type -> message.DeleteMessageRequest
+	14, // 20: message.MessageService.FetchConversationMessages:input_type -> message.FetchConversationMessagesRequest
+	16, // 21: message.MessageService.DeleteConversation:input_type -> message.DeleteConversationRequest
+	18, // 22: message.MessageService.FetchAllUserConversations:input_type -> message.FetchAllUserConversationsRequest
+	1,  // 23: message.MessageService.PingApiMessage:output_type -> message.PingApiMessageResponse
+	5,  // 24: message.MessageService.CustomerMessage:output_type -> message.CustomerMessageResponse
+	21, // 25: message.MessageService.ChatbotMessage:output_type -> message.ChatbotMessageResponse
+	3,  // 26: message.MessageService.PingTasksQueue:output_type -> message.PingTasksQueueResponse
+	23, // 27: message.MessageService.CreateTaskMessage:output_type -> message.CreateTaskMessageResponse
+	25, // 28: message.MessageService.CompleteTaskMessage:output_type -> message.CompleteTaskMessageResponse
+	7,  // 29: message.MessageService.UserSendMessage:output_type -> message.UserSendMessageResponse
+	9,  // 30: message.MessageService.EditMessage:output_type -> message.EditMessageResponse
+	11, // 31: message.MessageService.FetchMessage:output_type -> message.FetchMessageResponse
+	13, // 32: message.MessageService.DeleteMessage:output_type -> message.DeleteMessageResponse
+	15, // 33: message.MessageService.FetchConversationMessages:output_type -> message.FetchConversationMessagesResponse
+	17, // 34: message.MessageService.DeleteConversation:output_type -> message.DeleteConversationResponse
+	19, // 35: message.MessageService.FetchAllUserConversations:output_type -> message.FetchAllUserConversationsResponse
+	23, // [23:36] is the sub-list for method output_type
+	10, // [10:23] is the sub-list for method input_type
+	10, // [10:10] is the sub-list for extension type_name
+	10, // [10:10] is the sub-list for extension extendee
+	0,  // [0:10] is the sub-list for field type_name
 }
 
 func init() { file_grpc_proto_message_proto_init() }
@@ -2063,7 +2256,7 @@ func file_grpc_proto_message_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_grpc_proto_message_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*PingApiMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2075,7 +2268,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*PingApiMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2087,7 +2280,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*PingTasksQueueRequest); i {
 			case 0:
 				return &v.state
@@ -2099,7 +2292,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PingTasksQueueResponse); i {
 			case 0:
 				return &v.state
@@ -2111,7 +2304,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*CustomerMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2123,7 +2316,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*CustomerMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2135,7 +2328,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*UserSendMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2147,7 +2340,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*UserSendMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2159,7 +2352,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*EditMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2171,7 +2364,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*EditMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2183,7 +2376,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2195,7 +2388,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2207,7 +2400,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2219,7 +2412,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2231,7 +2424,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchConversationMessagesRequest); i {
 			case 0:
 				return &v.state
@@ -2243,7 +2436,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[15].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchConversationMessagesResponse); i {
 			case 0:
 				return &v.state
@@ -2255,7 +2448,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[16].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteConversationRequest); i {
 			case 0:
 				return &v.state
@@ -2267,7 +2460,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[17].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[17].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteConversationResponse); i {
 			case 0:
 				return &v.state
@@ -2279,7 +2472,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[18].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[18].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchAllUserConversationsRequest); i {
 			case 0:
 				return &v.state
@@ -2291,7 +2484,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[19].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[19].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchAllUserConversationsResponse); i {
 			case 0:
 				return &v.state
@@ -2303,7 +2496,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[20].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[20].Exporter = func(v any, i int) any {
 			switch v := v.(*ChatbotMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2315,7 +2508,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[21].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[21].Exporter = func(v any, i int) any {
 			switch v := v.(*ChatbotMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2327,7 +2520,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[22].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[22].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateTaskMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2339,7 +2532,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[23].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[23].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateTaskMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2351,7 +2544,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[24].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[24].Exporter = func(v any, i int) any {
 			switch v := v.(*CompleteTaskMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2363,7 +2556,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[25].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[25].Exporter = func(v any, i int) any {
 			switch v := v.(*CompleteTaskMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2375,7 +2568,31 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[26].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[26].Exporter = func(v any, i int) any {
+			switch v := v.(*ConversationRender); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_grpc_proto_message_proto_msgTypes[27].Exporter = func(v any, i int) any {
+			switch v := v.(*Participant); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_grpc_proto_message_proto_msgTypes[28].Exporter = func(v any, i int) any {
 			switch v := v.(*Conversation); i {
 			case 0:
 				return &v.state
@@ -2387,7 +2604,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[27].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[29].Exporter = func(v any, i int) any {
 			switch v := v.(*Message); i {
 			case 0:
 				return &v.state
@@ -2399,7 +2616,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[28].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[30].Exporter = func(v any, i int) any {
 			switch v := v.(*TaskData); i {
 			case 0:
 				return &v.state
@@ -2418,7 +2635,7 @@ func file_grpc_proto_message_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_grpc_proto_message_proto_rawDesc,
 			NumEnums:      0,
-			NumMessages:   29,
+			NumMessages:   31,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
