@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v3.20.3
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.27.3
 // source: grpc-proto/media.proto
 
 package grpc_proto
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	MediaService_PingApiMedia_FullMethodName    = "/media.MediaService/PingApiMedia"
@@ -46,8 +46,9 @@ func NewMediaServiceClient(cc grpc.ClientConnInterface) MediaServiceClient {
 }
 
 func (c *mediaServiceClient) PingApiMedia(ctx context.Context, in *PingApiMediaRequest, opts ...grpc.CallOption) (*PingApiMediaResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PingApiMediaResponse)
-	err := c.cc.Invoke(ctx, MediaService_PingApiMedia_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MediaService_PingApiMedia_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -55,8 +56,9 @@ func (c *mediaServiceClient) PingApiMedia(ctx context.Context, in *PingApiMediaR
 }
 
 func (c *mediaServiceClient) PostUploadMedia(ctx context.Context, in *PostUploadMediaRequest, opts ...grpc.CallOption) (*PostUploadMediaResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostUploadMediaResponse)
-	err := c.cc.Invoke(ctx, MediaService_PostUploadMedia_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MediaService_PostUploadMedia_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -64,8 +66,9 @@ func (c *mediaServiceClient) PostUploadMedia(ctx context.Context, in *PostUpload
 }
 
 func (c *mediaServiceClient) PostFetchMedia(ctx context.Context, in *PostFetchMediaRequest, opts ...grpc.CallOption) (*PostFetchMediaResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostFetchMediaResponse)
-	err := c.cc.Invoke(ctx, MediaService_PostFetchMedia_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MediaService_PostFetchMedia_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -73,8 +76,9 @@ func (c *mediaServiceClient) PostFetchMedia(ctx context.Context, in *PostFetchMe
 }
 
 func (c *mediaServiceClient) PostDeleteMedia(ctx context.Context, in *PostDeleteMediaRequest, opts ...grpc.CallOption) (*PostDeleteMediaResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostDeleteMediaResponse)
-	err := c.cc.Invoke(ctx, MediaService_PostDeleteMedia_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MediaService_PostDeleteMedia_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -82,8 +86,9 @@ func (c *mediaServiceClient) PostDeleteMedia(ctx context.Context, in *PostDelete
 }
 
 func (c *mediaServiceClient) PostsFetchMedia(ctx context.Context, in *PostsFetchMediaRequest, opts ...grpc.CallOption) (*PostsFetchMediaResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostsFetchMediaResponse)
-	err := c.cc.Invoke(ctx, MediaService_PostsFetchMedia_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MediaService_PostsFetchMedia_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -92,7 +97,7 @@ func (c *mediaServiceClient) PostsFetchMedia(ctx context.Context, in *PostsFetch
 
 // MediaServiceServer is the server API for MediaService service.
 // All implementations must embed UnimplementedMediaServiceServer
-// for forward compatibility
+// for forward compatibility.
 type MediaServiceServer interface {
 	PingApiMedia(context.Context, *PingApiMediaRequest) (*PingApiMediaResponse, error)
 	PostUploadMedia(context.Context, *PostUploadMediaRequest) (*PostUploadMediaResponse, error)
@@ -102,9 +107,12 @@ type MediaServiceServer interface {
 	mustEmbedUnimplementedMediaServiceServer()
 }
 
-// UnimplementedMediaServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedMediaServiceServer struct {
-}
+// UnimplementedMediaServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedMediaServiceServer struct{}
 
 func (UnimplementedMediaServiceServer) PingApiMedia(context.Context, *PingApiMediaRequest) (*PingApiMediaResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method PingApiMedia not implemented")
@@ -122,6 +130,7 @@ func (UnimplementedMediaServiceServer) PostsFetchMedia(context.Context, *PostsFe
 	return nil, status.Errorf(codes.Unimplemented, "method PostsFetchMedia not implemented")
 }
 func (UnimplementedMediaServiceServer) mustEmbedUnimplementedMediaServiceServer() {}
+func (UnimplementedMediaServiceServer) testEmbeddedByValue()                      {}
 
 // UnsafeMediaServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to MediaServiceServer will
@@ -131,6 +140,13 @@ type UnsafeMediaServiceServer interface {
 }
 
 func RegisterMediaServiceServer(s grpc.ServiceRegistrar, srv MediaServiceServer) {
+	// If the following call pancis, it indicates UnimplementedMediaServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&MediaService_ServiceDesc, srv)
 }
 
