@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v3.20.3
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.27.3
 // source: grpc-proto/social.proto
 
 package grpc_proto
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	SocialService_PingApiSocial_FullMethodName           = "/social.SocialService/PingApiSocial"
@@ -74,8 +74,9 @@ func NewSocialServiceClient(cc grpc.ClientConnInterface) SocialServiceClient {
 }
 
 func (c *socialServiceClient) PingApiSocial(ctx context.Context, in *PingApiSocialRequest, opts ...grpc.CallOption) (*PingApiSocialResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PingApiSocialResponse)
-	err := c.cc.Invoke(ctx, SocialService_PingApiSocial_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PingApiSocial_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -83,8 +84,9 @@ func (c *socialServiceClient) PingApiSocial(ctx context.Context, in *PingApiSoci
 }
 
 func (c *socialServiceClient) PostCreate(ctx context.Context, in *PostCreateRequest, opts ...grpc.CallOption) (*PostCreateResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostCreateResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostCreate_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostCreate_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -92,8 +94,9 @@ func (c *socialServiceClient) PostCreate(ctx context.Context, in *PostCreateRequ
 }
 
 func (c *socialServiceClient) PostsFetchByUserID(ctx context.Context, in *PostsFetchByUserIDRequest, opts ...grpc.CallOption) (*PostsFetchByUserIDResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostsFetchByUserIDResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostsFetchByUserID_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostsFetchByUserID_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -101,8 +104,9 @@ func (c *socialServiceClient) PostsFetchByUserID(ctx context.Context, in *PostsF
 }
 
 func (c *socialServiceClient) PostDelete(ctx context.Context, in *PostDeleteRequest, opts ...grpc.CallOption) (*PostDeleteResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostDeleteResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostDelete_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostDelete_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -110,8 +114,9 @@ func (c *socialServiceClient) PostDelete(ctx context.Context, in *PostDeleteRequ
 }
 
 func (c *socialServiceClient) PostsFetchVirtualLatest(ctx context.Context, in *PostsFetchVirtualLatestRequest, opts ...grpc.CallOption) (*PostsFetchVirtualLatestResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostsFetchVirtualLatestResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostsFetchVirtualLatest_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostsFetchVirtualLatest_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -119,8 +124,9 @@ func (c *socialServiceClient) PostsFetchVirtualLatest(ctx context.Context, in *P
 }
 
 func (c *socialServiceClient) PostsFetchByKeyword(ctx context.Context, in *PostsFetchByKeywordRequest, opts ...grpc.CallOption) (*PostsFetchByKeywordResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostsFetchByKeywordResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostsFetchByKeyword_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostsFetchByKeyword_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -128,8 +134,9 @@ func (c *socialServiceClient) PostsFetchByKeyword(ctx context.Context, in *Posts
 }
 
 func (c *socialServiceClient) PostsFetchByKeywords(ctx context.Context, in *PostsFetchByKeywordsRequest, opts ...grpc.CallOption) (*PostsFetchByKeywordsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostsFetchByKeywordsResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostsFetchByKeywords_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostsFetchByKeywords_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -137,8 +144,9 @@ func (c *socialServiceClient) PostsFetchByKeywords(ctx context.Context, in *Post
 }
 
 func (c *socialServiceClient) PostsFetchAllKeywords(ctx context.Context, in *PostsFetchAllKeywordsRequest, opts ...grpc.CallOption) (*PostsFetchAllKeywordsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostsFetchAllKeywordsResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostsFetchAllKeywords_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostsFetchAllKeywords_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -146,8 +154,9 @@ func (c *socialServiceClient) PostsFetchAllKeywords(ctx context.Context, in *Pos
 }
 
 func (c *socialServiceClient) PostToggleLike(ctx context.Context, in *PostToggleLikeRequest, opts ...grpc.CallOption) (*PostToggleLikeResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostToggleLikeResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostToggleLike_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostToggleLike_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -155,8 +164,9 @@ func (c *socialServiceClient) PostToggleLike(ctx context.Context, in *PostToggle
 }
 
 func (c *socialServiceClient) PostFetchLikes(ctx context.Context, in *PostFetchLikesRequest, opts ...grpc.CallOption) (*PostFetchLikesResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostFetchLikesResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostFetchLikes_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostFetchLikes_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -164,8 +174,9 @@ func (c *socialServiceClient) PostFetchLikes(ctx context.Context, in *PostFetchL
 }
 
 func (c *socialServiceClient) PostFetchViews(ctx context.Context, in *PostFetchViewsRequest, opts ...grpc.CallOption) (*PostFetchViewsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostFetchViewsResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostFetchViews_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostFetchViews_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -173,8 +184,9 @@ func (c *socialServiceClient) PostFetchViews(ctx context.Context, in *PostFetchV
 }
 
 func (c *socialServiceClient) PostCommentCreate(ctx context.Context, in *PostCommentCreateRequest, opts ...grpc.CallOption) (*PostCommentCreateResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostCommentCreateResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostCommentCreate_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostCommentCreate_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -182,8 +194,9 @@ func (c *socialServiceClient) PostCommentCreate(ctx context.Context, in *PostCom
 }
 
 func (c *socialServiceClient) PostCommentsFetch(ctx context.Context, in *PostCommentsFetchRequest, opts ...grpc.CallOption) (*PostCommentsFetchResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostCommentsFetchResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostCommentsFetch_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostCommentsFetch_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -191,8 +204,9 @@ func (c *socialServiceClient) PostCommentsFetch(ctx context.Context, in *PostCom
 }
 
 func (c *socialServiceClient) PostCommentDelete(ctx context.Context, in *PostCommentDeleteRequest, opts ...grpc.CallOption) (*PostCommentDeleteResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostCommentDeleteResponse)
-	err := c.cc.Invoke(ctx, SocialService_PostCommentDelete_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_PostCommentDelete_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -200,8 +214,9 @@ func (c *socialServiceClient) PostCommentDelete(ctx context.Context, in *PostCom
 }
 
 func (c *socialServiceClient) FollowCreate(ctx context.Context, in *FollowCreateRequest, opts ...grpc.CallOption) (*FollowCreateResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FollowCreateResponse)
-	err := c.cc.Invoke(ctx, SocialService_FollowCreate_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_FollowCreate_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -209,8 +224,9 @@ func (c *socialServiceClient) FollowCreate(ctx context.Context, in *FollowCreate
 }
 
 func (c *socialServiceClient) FollowDelete(ctx context.Context, in *FollowDeleteRequest, opts ...grpc.CallOption) (*FollowDeleteResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FollowDeleteResponse)
-	err := c.cc.Invoke(ctx, SocialService_FollowDelete_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_FollowDelete_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -218,8 +234,9 @@ func (c *socialServiceClient) FollowDelete(ctx context.Context, in *FollowDelete
 }
 
 func (c *socialServiceClient) FollowingsFetch(ctx context.Context, in *FollowingsFetchRequest, opts ...grpc.CallOption) (*FollowingsFetchResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FollowingsFetchResponse)
-	err := c.cc.Invoke(ctx, SocialService_FollowingsFetch_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_FollowingsFetch_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -227,8 +244,9 @@ func (c *socialServiceClient) FollowingsFetch(ctx context.Context, in *Following
 }
 
 func (c *socialServiceClient) FollowersFetch(ctx context.Context, in *FollowersFetchRequest, opts ...grpc.CallOption) (*FollowersFetchResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FollowersFetchResponse)
-	err := c.cc.Invoke(ctx, SocialService_FollowersFetch_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_FollowersFetch_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -236,8 +254,9 @@ func (c *socialServiceClient) FollowersFetch(ctx context.Context, in *FollowersF
 }
 
 func (c *socialServiceClient) FollowsCounterFetch(ctx context.Context, in *FollowsCounterFetchRequest, opts ...grpc.CallOption) (*FollowsCounterFetchResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FollowsCounterFetchResponse)
-	err := c.cc.Invoke(ctx, SocialService_FollowsCounterFetch_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, SocialService_FollowsCounterFetch_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -246,7 +265,7 @@ func (c *socialServiceClient) FollowsCounterFetch(ctx context.Context, in *Follo
 
 // SocialServiceServer is the server API for SocialService service.
 // All implementations must embed UnimplementedSocialServiceServer
-// for forward compatibility
+// for forward compatibility.
 type SocialServiceServer interface {
 	PingApiSocial(context.Context, *PingApiSocialRequest) (*PingApiSocialResponse, error)
 	PostCreate(context.Context, *PostCreateRequest) (*PostCreateResponse, error)
@@ -270,9 +289,12 @@ type SocialServiceServer interface {
 	mustEmbedUnimplementedSocialServiceServer()
 }
 
-// UnimplementedSocialServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedSocialServiceServer struct {
-}
+// UnimplementedSocialServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedSocialServiceServer struct{}
 
 func (UnimplementedSocialServiceServer) PingApiSocial(context.Context, *PingApiSocialRequest) (*PingApiSocialResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method PingApiSocial not implemented")
@@ -332,6 +354,7 @@ func (UnimplementedSocialServiceServer) FollowsCounterFetch(context.Context, *Fo
 	return nil, status.Errorf(codes.Unimplemented, "method FollowsCounterFetch not implemented")
 }
 func (UnimplementedSocialServiceServer) mustEmbedUnimplementedSocialServiceServer() {}
+func (UnimplementedSocialServiceServer) testEmbeddedByValue()                       {}
 
 // UnsafeSocialServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to SocialServiceServer will
@@ -341,6 +364,13 @@ type UnsafeSocialServiceServer interface {
 }
 
 func RegisterSocialServiceServer(s grpc.ServiceRegistrar, srv SocialServiceServer) {
+	// If the following call pancis, it indicates UnimplementedSocialServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&SocialService_ServiceDesc, srv)
 }
 
